JRA-18 Refactored home page to show all results on it
