+++ app/src/test/java/com/jalasoft/routesapp/ui/home/HomeViewModelTest.kt
@@ -1,6 +1,11 @@
 package com.jalasoft.routesapp.ui.home
 
+import com.google.maps.android.PolyUtil
+import com.jalasoft.routesapp.data.api.models.gmaps.EndLocation
+import com.jalasoft.routesapp.data.api.models.gmaps.StartLocation
+import com.jalasoft.routesapp.data.source.FakeDirectionsManager
 import com.jalasoft.routesapp.data.source.FakePlaceManager
+import com.jalasoft.routesapp.data.source.FakeRoutesData
 import com.jalasoft.routesapp.ui.home.viewModel.HomeViewModel
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
@@ -24,13 +29,15 @@ class HomeViewModelTest : TestCase() {
 
     private lateinit var viewModel: HomeViewModel
     private lateinit var fakeManager: FakePlaceManager
+    private lateinit var fakeDirectionsManager: FakeDirectionsManager
 
     @Before
     public override fun setUp() {
         super.setUp()
         hiltRule.inject()
         fakeManager = FakePlaceManager()
-        viewModel = HomeViewModel(fakeManager)
+        fakeDirectionsManager = FakeDirectionsManager()
+        viewModel = HomeViewModel(fakeManager, fakeDirectionsManager)
     }
 
     @Test
@@ -43,5 +50,17 @@ class HomeViewModelTest : TestCase() {
     fun `When the user set the destination location the value should update`() {
         viewModel.setDestination(PlaceManagerFakeData.fakeLatLngDestination)
         assertEquals(viewModel.selectedDestination.value, PlaceManagerFakeData.fakeLatLngDestination)
+    }
+
+    @Test
+    fun `Given two stops, when they belong to different transportation categories, then returns all the points to join this stops`() {
+        val startLocation = StartLocation(-16.5244779,	-68.1253892)
+        val endLocation = EndLocation(-16.5255314, -68.1254204)
+        val points = viewModel.fetchDirections(startLocation, endLocation)?.first()?.overviewPolyline?.points
+        points?.let {
+            assertEquals(PolyUtil.decode(it), FakeRoutesData.directionsPointLst)
+        } ?: run {
+            assertEquals(false, true)
+        }
     }
 }
