+++ app/src/test/java/com/jalasoft/routesapp/ui/routes/viewModel/RoutesViewModelTest.kt
@@ -2,10 +2,6 @@ package com.jalasoft.routesapp.ui.routes.viewModel
 
 import android.content.Context
 import androidx.test.platform.app.InstrumentationRegistry
-import com.google.maps.android.PolyUtil
-import com.jalasoft.routesapp.data.api.models.gmaps.EndLocation
-import com.jalasoft.routesapp.data.api.models.gmaps.StartLocation
-import com.jalasoft.routesapp.data.source.FakeDirectionsManager
 import com.jalasoft.routesapp.data.source.FakeRoutesData
 import com.jalasoft.routesapp.data.source.FakeRoutesManager
 import dagger.hilt.android.testing.HiltAndroidRule
@@ -30,7 +26,6 @@ class RoutesViewModelTest : TestCase() {
 
     private lateinit var viewModel: RoutesViewModel
     private lateinit var fakeManager: FakeRoutesManager
-    private lateinit var fakeDirectionsManager: FakeDirectionsManager
     private lateinit var context: Context
 
     @Before
@@ -38,40 +33,27 @@ class RoutesViewModelTest : TestCase() {
         super.setUp()
         hiltRule.inject()
         fakeManager = FakeRoutesManager()
-        fakeDirectionsManager = FakeDirectionsManager()
-        viewModel = RoutesViewModel(fakeManager, fakeDirectionsManager)
+        viewModel = RoutesViewModel(fakeManager)
         context = InstrumentationRegistry.getInstrumentation().context
     }
 
     @Test
     fun `Given the first call for fetching the lines, when the line screen appears, then returns all the Lines from the current city`() {
         viewModel.fetchLines(context)
-        assertEquals(viewModel._routesList.value, FakeRoutesData.lineInfo)
+        assertEquals(viewModel.routesList.value, FakeRoutesData.lineInfo)
     }
 
     @Test
     fun `Given the first call for fetching lines, when there is a filter criteria, then returns all the Lines that match the filter criteria`() {
         viewModel.fetchLines(context)
         viewModel.filterLines("Line")
-        assertEquals(viewModel._routesList.value, FakeRoutesData.lineInfo.filter { it.name.contains("Line") })
+        assertEquals(viewModel.routesList.value, FakeRoutesData.lineInfo.filter { it.name.contains("Line") })
     }
 
     @Test
     fun `Given the first call for fetching lines, when there is a filter criteria that does not match, then returns an empty result`() {
         viewModel.fetchLines(context)
         viewModel.filterLines("5")
-        assertTrue(viewModel._routesList.value!!.isEmpty())
-    }
-
-    @Test
-    fun `Given two stops, when they belong to different transportation categories, then returns all the points to join this stops`() {
-        val startLocation = StartLocation(-16.5244779,	-68.1253892)
-        val endLocation = EndLocation(-16.5255314, -68.1254204)
-        val points = viewModel.fetchDirections(startLocation, endLocation)?.first()?.overviewPolyline?.points
-        points?.let {
-            assertEquals(PolyUtil.decode(it), FakeRoutesData.directionsPointLst)
-        } ?: run {
-            assertEquals(false, true)
-        }
+        assertTrue(viewModel.routesList.value!!.isEmpty())
     }
 }
